Convert selected columns to factor type

This function takes a data frame and converts specified columns to factors.

Usage

convert_columns_to_factor(df)

Arguments

df

A data frame containing the columns: room_type, host_is_superhost, city, weekdays

Value

A data frame with the specified columns converted to factors

Examples

df <- data.frame(
  room_type = c("Private room", "Entire home/apt"),
  host_is_superhost = c("t", "f"),
  city = c("Madrid", "Barcelona"),
  weekdays = c(TRUE, FALSE),
  price = c(100, 150)
)
convert_columns_to_factor(df)
#>         room_type host_is_superhost      city weekdays price
#> 1    Private room                 t    Madrid     TRUE   100
#> 2 Entire home/apt                 f Barcelona    FALSE   150